f1-pipeline/requirements.txt
# ── Core dependencies ────────────────────────────────
boto3>=1.35.0
kaggle>=2.0.0
python-dotenv>=1.0.0
